personel listeleme ekranı yapıldı

+++ Entities/Concrete/HardKod.cs
@@ -1,14 +1,11 @@
 ﻿using Core.Entities;
 using Entities.Base;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Entities.Concrete
 {
     public class HardKod : EntityBase, IEntity
     {
-        public long UstKodId { get; set; }
+        public int UstKodId { get; set; }
         public string Ad { get; set; }
     }
 }
